Do not require config setup for scenarios
This commit removes the hard requirement of
having the config file specified with a scenario.

This is to allow for the case where this node
terminus is used, but the environment has not
been fully set up yet. For my specific case, this
is when I run an initial puppet apply command to set
up the base environment.