use hostname
Site manifests in puppet will try not
only the certname to detect matches for
classification, but it also tries to remove
each domain name section when checking for a match.

This commit adds the same logic to the scenario
node terminus so that it will act as similar as
possible to site manifests.