Allow hiera overrides for global params
This patch allows for the global param variables
(previously configured in config.yaml) to be
specified in its own hierarchy directory
(data/global_hiera_params). This code was added
b/c we realized that that many of the global configs
had different default values based on the scenario
(ie db_type is galera for the HA scenario).

This patch allows you to specify a hierarchy for
global configs. It allows you to use the scenario
in the override hierarchy, but nothing else (so
it currently uses the scenario and all staic categories).