introduce cert_for_facts option
The previous implementation of fact retrieval for
a different machine that what you were running on
was a but naive.

It tried to use Puppet[:certname] for this purpose.
This seemed to work for most use cases, but failed
for the puppetdb facts terminus b/c this terminus
uses the certname to find the certificates to use
for ssl auth.

This commit adds its own option cert_for_facts that
can be used to specify an external cert name to
lookup.